Fix app state to commit properly
